Minor rework of unuathenticated dashboard
